update to main

+++ code/game/machinery/OpTable.dm
@@ -5,16 +5,18 @@
 	icon_state = "table2-idle"
 	density = 1
 	anchored = 1.0
-	use_power = 1
+	use_power = IDLE_POWER_USE
 	idle_power_usage = 1
 	active_power_usage = 5
 	var/mob/living/carbon/human/victim = null
 	var/strapped = 0.0
 
 	var/obj/machinery/computer/operating/computer = null
-	buckle_lying = 90
+	buckle_lying = -1
 	var/no_icon_updates = 0 //set this to 1 if you don't want the icons ever changing
 	var/list/injected_reagents = list()
+	var/reagent_target_amount = 1
+	var/inject_amount = 1
 
 /obj/machinery/optable/New()
 	..()
@@ -33,27 +35,6 @@
 		victim = null
 	return ..()
 
-/obj/machinery/optable/ex_act(severity)
-	switch(severity)
-		if(1.0)
-			//SN src = null
-			qdel(src)
-			return
-		if(2.0)
-			if(prob(50))
-				//SN src = null
-				qdel(src)
-				return
-		if(3.0)
-			if(prob(25))
-				src.density = 0
-		else
-	return
-
-/obj/machinery/optable/blob_act()
-	if(prob(75))
-		qdel(src)
-
 /obj/machinery/optable/attack_hulk(mob/living/carbon/human/user, does_attack_animation = FALSE)
 	if(user.a_intent == INTENT_HARM)
 		..(user, TRUE)
@@ -71,7 +52,7 @@
 
 
 /obj/machinery/optable/MouseDrop_T(atom/movable/O as mob|obj, mob/user as mob)
-	if(usr.stat || (!ishuman(user) && !isrobot(user)) || user.restrained() || !check_table(user) || user.weakened || user.stunned)
+	if(usr.stat || (!ishuman(user) && !isrobot(user)) || user.restrained() || !check_table(user) || user.IsWeakened() || user.stunned)
 		return
 
 	if(!ismob(O)) //humans only
@@ -97,7 +78,7 @@
 		icon_state = "table2-idle"
 	return 0
 
-/obj/machinery/optable/Crossed(atom/movable/AM)
+/obj/machinery/optable/Crossed(atom/movable/AM, oldloc)
 	. = ..()
 	if(iscarbon(AM) && LAZYLEN(injected_reagents))
 		to_chat(AM, "<span class='danger'>You feel a series of tiny pricks!</span>")
@@ -106,9 +87,9 @@
 	check_victim()
 	if(LAZYLEN(injected_reagents))
 		for(var/mob/living/carbon/C in get_turf(src))
+			var/datum/reagents/R = C.reagents
 			for(var/chemical in injected_reagents)
-				if(C.reagents.get_reagent_amount(chemical) < 1)
-					C.reagents.add_reagent(chemical, 1)
+				R.check_and_add(chemical,reagent_target_amount,inject_amount)
 
 /obj/machinery/optable/proc/take_victim(mob/living/carbon/C, mob/living/carbon/user as mob)
 	if(C == user)
@@ -120,6 +101,8 @@
 	C.forceMove(loc)
 	if(user.pulling == C)
 		user.stop_pulling()
+	if(C.s_active) //Close the container opened
+		C.s_active.close(C)
 	for(var/obj/O in src)
 		O.loc = src.loc
 	src.add_fingerprint(user)
@@ -142,19 +125,23 @@
 
 	take_victim(usr,usr)
 
-/obj/machinery/optable/attackby(obj/item/W as obj, mob/living/carbon/user as mob, params)
-	if(istype(W, /obj/item/grab))
-		if(iscarbon(W:affecting))
-			take_victim(W:affecting,usr)
-			qdel(W)
-			return
-	if(istype(W, /obj/item/wrench))
-		playsound(src.loc, W.usesound, 50, 1)
-		if(do_after(user, 20 * W.toolspeed, target = src))
-			to_chat(user, "<span class='notice'>You deconstruct the table.</span>")
-			new /obj/item/stack/sheet/plasteel(loc, 5)
-			qdel(src)
+/obj/machinery/optable/attackby(obj/item/I, mob/living/carbon/user, params)
+	if(istype(I, /obj/item/grab))
+		var/obj/item/grab/G = I
+		if(iscarbon(G.affecting))
+			take_victim(G.affecting, user)
+			qdel(G)
+	else
+		return ..()
 
+/obj/machinery/optable/wrench_act(mob/user, obj/item/I)
+	. = TRUE
+	if(!I.tool_start_check(src, user, 0))
+		return
+	if(I.use_tool(src, user, 20, volume = I.tool_volume))
+		to_chat(user, "<span class='notice'>You deconstruct the table.</span>")
+		new /obj/item/stack/sheet/plasteel(loc, 5)
+		qdel(src)
 
 /obj/machinery/optable/proc/check_table(mob/living/carbon/patient as mob)
 	if(src.victim && get_turf(victim) == get_turf(src) && victim.lying)
